pyproject.toml
[tool.poetry]
name = "astro-gen-backend"
version = "0.1.0"
description = "Backend for Astro Natal Chart Generator"
authors = ["Your Name <email@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
pydantic = "^2.10.0"
pydantic-settings = "^2.7.0"
sqlalchemy = "^2.0.36"
greenlet = "^3.1.0"
asyncpg = "^0.30.0"
alembic = "^1.14.0"
redis = "^5.2.0"
celery = "^5.4.0"
httpx = "^0.28.0"
httpx-oauth = "^0.16.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
PyJWT = {extras = ["crypto"], version = "^2.10.0"}
cryptography = "^44.0.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "^4.2.0"
python-multipart = "^0.0.20"
itsdangerous = "^2.2.0"
yookassa = "^3.3.0"
kerykeion = "^4.26.0"
cairosvg = "^2.7.0"
weasyprint = "^63.0"
markdown = "^3.7.0"
jinja2 = "^3.1.0"
fastapi-mail = "^1.4.0"
sqladmin = "^0.20.0"
slowapi = "^0.1.9"
openai = "^1.59.0"
structlog = "^25.0.0"
tenacity = "^9.0.0"
email-validator = "^2.2.0"
pytz = "^2024.2"
aiofiles = "^24.0.0"
aiosqlite = "^0.20.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-asyncio = "^0.23.6"
httpx = "^0.27.0"
fakeredis = "^2.26.0"
faker = "^24.4.0"
pytest-cov = "^5.0.0"
black = "^24.3.0"
isort = "^5.13.2"
flake8 = "^7.0.0"
mypy = "^1.9.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
pythonpath = ["."]
testpaths = ["tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"